+++ core/src/test/java/org/apache/calcite/rel/rules/SortRemoveRuleTest.java
@@ -85,7 +85,7 @@ public final class SortRemoveRuleTest {
    * <p>Since join inputs are sorted, and this join preserves the order of the
    * left input, there shouldn't be any sort operator above the join.
    */
-  @Test public void removeSortOverEnumerableJoin() throws Exception {
+  @Test public void removeSortOverEnumerableHashJoin() throws Exception {
     RuleSet prepareRules =
         RuleSets.ofList(
             SortProjectTransposeRule.INSTANCE,
@@ -103,7 +103,7 @@ public final class SortRemoveRuleTest {
       assertThat(
           toString(actualPlan),
           allOf(
-              containsString("EnumerableJoin"),
+              containsString("EnumerableHashJoin"),
               not(containsString("EnumerableSort"))));
     }
   }
@@ -116,7 +116,7 @@ public final class SortRemoveRuleTest {
    * <p>Since join inputs are sorted, and this join preserves the order of the
    * left input, there shouldn't be any sort operator above the join.
    */
-  @Test public void removeSortOverEnumerableThetaJoin() throws Exception {
+  @Test public void removeSortOverEnumerableNestedLoopJoin() throws Exception {
     RuleSet prepareRules =
         RuleSets.ofList(
             SortProjectTransposeRule.INSTANCE,
@@ -124,8 +124,8 @@ public final class SortRemoveRuleTest {
             EnumerableRules.ENUMERABLE_PROJECT_RULE,
             EnumerableRules.ENUMERABLE_SORT_RULE,
             EnumerableRules.ENUMERABLE_TABLE_SCAN_RULE);
-    // Inner join is not considered since the ENUMERABLE_JOIN_RULE does not generate a theta join
-    // in the case of inner joins.
+    // Inner join is not considered since the ENUMERABLE_JOIN_RULE does not generate a nestedLoop
+    // join in the case of inner joins.
     for (String joinType : Arrays.asList("left", "right", "full")) {
       String sql =
           "select e.\"deptno\" from \"hr\".\"emps\" e "
@@ -136,7 +136,7 @@ public final class SortRemoveRuleTest {
       assertThat(
           toString(actualPlan),
           allOf(
-              containsString("EnumerableThetaJoin"),
+              containsString("EnumerableNestedLoopJoin"),
               not(containsString("EnumerableSort"))));
     }
   }
@@ -147,13 +147,14 @@ public final class SortRemoveRuleTest {
    *
    * <p>Since join inputs are sorted, and this join preserves the order of the
    * left input, there shouldn't be any sort operator above the join.
+   *
+   * <p>Until CALCITE-2018 is fixed we can add back EnumerableRules.ENUMERABLE_SORT_RULE
    */
   @Test public void removeSortOverEnumerableCorrelate() throws Exception {
     RuleSet prepareRules =
         RuleSets.ofList(
             SortProjectTransposeRule.INSTANCE,
-            JoinToCorrelateRule.JOIN,
-            EnumerableRules.ENUMERABLE_SORT_RULE,
+            JoinToCorrelateRule.INSTANCE,
             EnumerableRules.ENUMERABLE_PROJECT_RULE,
             EnumerableRules.ENUMERABLE_CORRELATE_RULE,
             EnumerableRules.ENUMERABLE_FILTER_RULE,
@@ -188,7 +189,7 @@ public final class SortRemoveRuleTest {
             SemiJoinRule.JOIN,
             EnumerableRules.ENUMERABLE_PROJECT_RULE,
             EnumerableRules.ENUMERABLE_SORT_RULE,
-            EnumerableRules.ENUMERABLE_SEMI_JOIN_RULE,
+            EnumerableRules.ENUMERABLE_JOIN_RULE,
             EnumerableRules.ENUMERABLE_FILTER_RULE,
             EnumerableRules.ENUMERABLE_TABLE_SCAN_RULE);
     String sql =
@@ -199,7 +200,7 @@ public final class SortRemoveRuleTest {
     assertThat(
         toString(actualPlan),
         allOf(
-            containsString("EnumerableSemiJoin"),
+            containsString("EnumerableHashJoin"),
             not(containsString("EnumerableSort"))));
   }