+++ core/src/test/java/org/apache/calcite/test/MaterializationTest.java
@@ -1614,7 +1614,7 @@ public class MaterializationTest {
         HR_FKUK_MODEL,
         CalciteAssert.checkResultContains(
             "EnumerableAggregate(group=[{0}], S=[$SUM0($2)])\n"
-                + "  EnumerableJoin(condition=[=($1, $3)], joinType=[inner])\n"
+                + "  EnumerableHashJoin(condition=[=($1, $3)], joinType=[inner])\n"
                 + "    EnumerableTableScan(table=[[hr, m0]])\n"
                 + "    EnumerableTableScan(table=[[hr, depts]])"));
   }
@@ -1633,7 +1633,7 @@ public class MaterializationTest {
         HR_FKUK_MODEL,
         CalciteAssert.checkResultContains(
             "EnumerableAggregate(group=[{4}], S=[$SUM0($2)])\n"
-                + "  EnumerableJoin(condition=[=($1, $3)], joinType=[inner])\n"
+                + "  EnumerableHashJoin(condition=[=($1, $3)], joinType=[inner])\n"
                 + "    EnumerableTableScan(table=[[hr, m0]])\n"
                 + "    EnumerableTableScan(table=[[hr, depts]])"));
   }
@@ -1795,7 +1795,7 @@ public class MaterializationTest {
         HR_FKUK_MODEL,
         CalciteAssert.checkResultContains(
             "EnumerableCalc(expr#0..2=[{inputs}], empid=[$t1])\n"
-                + "  EnumerableJoin(condition=[=($0, $2)], joinType=[inner])\n"
+                + "  EnumerableHashJoin(condition=[=($0, $2)], joinType=[inner])\n"
                 + "    EnumerableCalc(expr#0=[{inputs}], expr#1=[CAST($t0):VARCHAR], name=[$t1])\n"
                 + "      EnumerableTableScan(table=[[hr, m0]])\n"
                 + "    EnumerableCalc(expr#0..1=[{inputs}], expr#2=[CAST($t1):VARCHAR], empid=[$t0], name0=[$t2])\n"
@@ -1814,7 +1814,7 @@ public class MaterializationTest {
         HR_FKUK_MODEL,
         CalciteAssert.checkResultContains(
             "EnumerableCalc(expr#0..4=[{inputs}], empid=[$t2])\n"
-                + "  EnumerableJoin(condition=[=($1, $4)], joinType=[inner])\n"
+                + "  EnumerableHashJoin(condition=[=($1, $4)], joinType=[inner])\n"
                 + "    EnumerableCalc(expr#0=[{inputs}], expr#1=[CAST($t0):VARCHAR], proj#0..1=[{exprs}])\n"
                 + "      EnumerableTableScan(table=[[hr, m0]])\n"
                 + "    EnumerableCalc(expr#0..1=[{inputs}], expr#2=[CAST($t1):VARCHAR], proj#0..2=[{exprs}])\n"
@@ -1855,12 +1855,14 @@ public class MaterializationTest {
   }
 
   @Test public void testJoinMaterialization11() {
-    checkNoMaterialize(
+    checkMaterialize(
         "select \"empid\" from \"emps\"\n"
             + "join \"depts\" using (\"deptno\")",
         "select \"empid\" from \"emps\"\n"
             + "where \"deptno\" in (select \"deptno\" from \"depts\")",
-        HR_FKUK_MODEL);
+        HR_FKUK_MODEL,
+        CalciteAssert.checkResultContains(
+            "PLAN=EnumerableTableScan(table=[[hr, m0]])"));
   }
 
   @Test public void testJoinMaterialization12() {